+++ software/open-qm-base-station/src/main/java/tech/ebp/oqm/baseStation/service/mongo/file/MongoFileService.java
@@ -12,6 +12,7 @@ import com.mongodb.client.gridfs.model.GridFSFile;
 import com.mongodb.client.gridfs.model.GridFSUploadOptions;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Sorts;
+import jakarta.annotation.PostConstruct;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -28,6 +29,7 @@ import org.bson.codecs.EncoderContext;
 import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
 import tech.ebp.oqm.baseStation.model.object.FileMainObject;
+import tech.ebp.oqm.baseStation.model.object.media.FileHashes;
 import tech.ebp.oqm.baseStation.model.object.media.FileMetadata;
 import tech.ebp.oqm.baseStation.rest.search.SearchObject;
 import tech.ebp.oqm.baseStation.service.TempFileService;
@@ -35,9 +37,11 @@ import tech.ebp.oqm.baseStation.service.mongo.MongoObjectService;
 import tech.ebp.oqm.baseStation.service.mongo.MongoService;
 import tech.ebp.oqm.baseStation.service.mongo.utils.FileContentsGet;
 
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -72,6 +76,28 @@ public abstract class MongoFileService<T extends FileMainObject, S extends Searc
 		super(objectMapper, mongoClient, database, clazz);
 		this.fileMetadataCodec = this.getDatabase().getCodecRegistry().get(FileMetadata.class);
 		this.tempFileService = tempFileService;
+	}
+	
+	@PostConstruct
+	public void setup(){
+		// should probably be a TODO to remove this, but unsure how we ever might be able to.
+		//ensure gridfs bucket storage is initialized. Required to avoid trying to create during a transaction, which is unsupported by Mongodb.
+		if(this.getGridFSBucket().find().limit(1).first() == null){
+			FileMetadata metadata = new FileMetadata(
+				"disregard_init_file_disregard",
+				0,
+				FileHashes.builder().md5("").sha1("").sha256("").build(),
+				FileMetadata.TIKA.detect("plain.txt"),
+				ZonedDateTime.now()
+			);
+			
+			GridFSUploadOptions ops = this.getUploadOps(metadata);
+			GridFSBucket bucket = this.getGridFSBucket();
+			String filename = "init";
+			
+			ObjectId id = bucket.uploadFromStream(filename, new ByteArrayInputStream("".getBytes()), ops);
+			bucket.delete(id);
+		}
 	}
 	
 	@Override
